requirements.txt
contourpy==1.3.2 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
fast-simplification==0.1.13 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.61.1 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.9 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.10.8 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.2.6 ; python_version >= "3.10" and python_version < "4.0"
packaging==26.0 ; python_version >= "3.10" and python_version < "4.0"
pillow==12.1.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
trimesh==4.11.1 ; python_version >= "3.10" and python_version < "4.0"
